Extracted divi page builder specific code to separate file
